requirements.txt
bottle>=0.12.7
paste>=1.7.5.1
MySQL-Python>=1.2.5
SQLAlchemy>=0.9.7
bottle_sqlalchemy>=0.4.1
httplib2>=0.9
python-novaclient>=2.18.1
